@jkeiser 2020-07-28
On Windows, twitter runs stage 2 a little bit faster, at 4.6744 GB/s before and 4.7554 GB/s after. These results translate to about a 1% overall difference.